requirements.txt
pandas>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
scikit-learn>=1.3
scipy>=1.11
numpy>=1.24
